requirements.txt
tensorflow-gpu==1.12.0
numpy==1.13
scipy==0.19.0
soundfile==0.9
tqdm==4.11.0
Cython==0.25.2
h5py==2.7.0